requirements.txt
TA-Lib
pandas-datareader
PyYAML
ipykernel
